Edits
I reduced the joy/pain image file size to 234 kb. I changed the button
color and text. I fixed the html because I added the images in the right
column to the main folder where the index is, but I had the html lead to
the same images in the images folder. I think that should fix it, but
let me know.